# Conflicts:
#	src/main/java/uz/elmurodov/ui/task/TaskUI.java

+++ src/main/java/uz/elmurodov/dtos/task/TaskUpdateDto.java
@@ -13,7 +13,7 @@ public class TaskUpdateDto extends GenericDto {
 	private int level;
 
 	@SerializedName("project_column_id")
-	private int projectColumnId;
+	private Long projectColumnId;
 
 	private String name;
